keep all functions of the same struct in the same file
