+++ core/pkg/server/handler.go
@@ -3,9 +3,12 @@ package server
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
+	"strconv"
+	"strings"
 
 	"github.com/wandb/wandb/core/pkg/monitor"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -164,9 +167,8 @@ func NewHandler(
 // Do starts the handler
 func (h *Handler) Do(inChan <-chan *service.Record) {
 	defer h.logger.Reraise()
-	h.logger.Info("handler: started", "stream_id", h.settings.RunId)
+	h.logger.Info("handler started", "stream_id", h.settings.RunId)
 	for record := range inChan {
-		h.logger.Debug("handling record", "record", record)
 		h.handleRecord(record)
 	}
 	h.Close()
@@ -175,7 +177,7 @@ func (h *Handler) Do(inChan <-chan *service.Record) {
 func (h *Handler) Close() {
 	close(h.outChan)
 	close(h.fwdChan)
-	h.logger.Debug("handler: closed", "stream_id", h.settings.RunId)
+	h.logger.Debug("handler closed", "stream_id", h.settings.RunId)
 }
 
 func (h *Handler) sendResponse(record *service.Record, response *service.Response) {
@@ -213,8 +215,7 @@ func (h *Handler) sendRecord(record *service.Record) {
 //gocyclo:ignore
 func (h *Handler) handleRecord(record *service.Record) {
 	h.summaryHandler.Debounce(h.sendSummary)
-	recordType := record.GetRecordType()
-	h.logger.Debug("handle: got a message", "record_type", recordType)
+	h.logger.Debug("handler handles a record", "record", record)
 	switch x := record.RecordType.(type) {
 	case *service.Record_Alert:
 		h.handleAlert(record)
@@ -257,7 +258,7 @@ func (h *Handler) handleRecord(record *service.Record) {
 	case *service.Record_UseArtifact:
 		h.handleUseArtifact(record)
 	case nil:
-		err := fmt.Errorf("handleRecord: record type is nil")
+		err := fmt.Errorf("handleRecord: nil record type")
 		h.logger.CaptureFatalAndPanic("error handling record", err)
 	default:
 		err := fmt.Errorf("handleRecord: unknown record type %T", x)
@@ -321,6 +322,9 @@ func (h *Handler) handleRequest(record *service.Record) {
 	case *service.Request_SenderRead:
 		h.handleSenderRead(record)
 		response = nil
+	case nil:
+		err := fmt.Errorf("handleRequest: nil request type")
+		h.logger.CaptureFatalAndPanic("error handling request", err)
 	default:
 		err := fmt.Errorf("handleRequest: unknown request type %T", x)
 		h.logger.CaptureFatalAndPanic("error handling request", err)
@@ -927,4 +931,355 @@ func (h *Handler) handleTBrecord(record *service.Record) {
 
 func (h *Handler) GetRun() *service.RunRecord {
 	return h.runRecord
+}
+
+func (h *Handler) sampleHistory(history *service.HistoryRecord) {
+	var value float32
+	if h.sampledHistory == nil {
+		h.sampledHistory = make(map[string]*ReservoirSampling[float32])
+	}
+	for _, item := range history.GetItem() {
+		err := json.Unmarshal([]byte(item.ValueJson), &value)
+		if err != nil {
+			continue
+		}
+		if _, ok := h.sampledHistory[item.Key]; !ok {
+			h.sampledHistory[item.Key] = &ReservoirSampling[float32]{
+				k:     48,
+				delta: 0.0005,
+			}
+		}
+		h.sampledHistory[item.Key].Add(value)
+	}
+}
+
+// flushHistory flushes a history record. It is responsible for handling the history record internally,
+// processing it, and forwarding it to the Writer.
+func (h *Handler) flushHistory(history *service.HistoryRecord) {
+	if history.GetItem() == nil {
+		return
+	}
+
+	// adds internal history items to the history record
+	// these items are used for internal bookkeeping and are not sent by the user
+	// TODO: add a timestamp field to the history record
+	var runTime float64 = 0
+	if item, ok := h.activeHistory.GetItem("_timestamp"); ok {
+		value := item.GetValueJson()
+		val, err := strconv.ParseFloat(value, 64)
+		if err != nil {
+			h.logger.CaptureError("error parsing timestamp", err)
+		} else {
+			runTime = val - h.timer.GetStartTimeMicro()
+		}
+	}
+	history.Item = append(history.Item,
+		&service.HistoryItem{Key: "_runtime", ValueJson: fmt.Sprintf("%f", runTime)},
+	)
+	if !h.settings.GetXShared().GetValue() {
+		history.Item = append(history.Item,
+			&service.HistoryItem{Key: "_step", ValueJson: fmt.Sprintf("%d", history.GetStep().GetNum())},
+		)
+	}
+
+	// handles all history items. It is responsible for matching current history
+	// items with defined metrics, and creating new metrics if needed. It also handles step metric in case
+	// it needs to be synced, but not part of the history record.
+	// This means that there are metrics defined for this run
+	if h.metricHandler != nil {
+		for _, item := range history.GetItem() {
+			step := h.imputeStepMetric(item)
+			// TODO: fix this, we update history while we are iterating over it
+			// TODO: handle nested step metrics (e.g. step defined by another step)
+			if step != nil {
+				history.Item = append(history.Item, step)
+			}
+		}
+	}
+
+	h.sampleHistory(history)
+
+	record := &service.Record{
+		RecordType: &service.Record_History{History: history},
+	}
+	h.sendRecord(record)
+
+	// TODO unify with handleSummary
+	// TODO add an option to disable summary (this could be quite expensive)
+	if h.summaryHandler == nil {
+		return
+	}
+	summary := corelib.ConsolidateSummaryItems(h.summaryHandler.consolidatedSummary, history.GetItem())
+	h.summaryHandler.updateSummaryDelta(summary)
+}
+
+// handleHistory handles a history record. This is the main entry point for history records.
+// It is responsible for handling the history record internally, processing it,
+// and forwarding it to the Writer.
+func (h *Handler) handleHistory(history *service.HistoryRecord) {
+	// TODO replace history encoding with a map, this will make it easier to handle history
+	h.activeHistory = NewActiveHistory(
+		WithStep(history.GetStep().GetNum()),
+	)
+	h.activeHistory.UpdateValues(history.GetItem())
+
+	h.flushHistory(history)
+
+	h.activeHistory.Flush()
+}
+
+// imputeStepMetric imputes a step metric if it needs to be synced, but not part of the history record.
+func (h *Handler) imputeStepMetric(item *service.HistoryItem) *service.HistoryItem {
+
+	// check if history item matches a defined metric or a glob metric
+	metric := h.matchHistoryItemMetric(item)
+
+	key := metric.GetStepMetric()
+	// check if step metric is defined and if it needs to be synced
+	if !(metric.GetOptions().GetStepSync() && key != "") {
+		return nil
+	}
+
+	// check if step metric is already in history
+	if _, ok := h.activeHistory.GetItem(key); ok {
+		return nil
+	}
+
+	// we use the summary value of the metric as the algorithm for imputing the step metric
+	if value, ok := h.summaryHandler.consolidatedSummary[key]; ok {
+		// TODO: add nested key support
+		hi := &service.HistoryItem{
+			Key:       key,
+			ValueJson: value,
+		}
+		h.activeHistory.UpdateValues([]*service.HistoryItem{hi})
+		return hi
+	}
+	return nil
+}
+
+// matchHistoryItemMetric matches a history item with a defined metric or creates a new metric if needed.
+func (h *Handler) matchHistoryItemMetric(item *service.HistoryItem) *service.MetricRecord {
+
+	// ignore internal history items
+	if strings.HasPrefix(item.Key, "_") {
+		return nil
+	}
+
+	// check if history item matches a defined metric exactly, if it does return the metric
+	if metric, ok := h.metricHandler.definedMetrics[item.Key]; ok {
+		return metric
+	}
+
+	// if a new metric was created, we need to handle it
+	metric := h.metricHandler.createMatchingGlobMetric(item.Key)
+	if metric != nil {
+		record := &service.Record{
+			RecordType: &service.Record_Metric{
+				Metric: metric,
+			},
+			Control: &service.Control{
+				Local: true,
+			},
+		}
+		h.handleMetric(record, metric)
+	}
+	return metric
+}
+
+// handlePartialHistory handles a partial history request. Collects the history items until a full
+// history record is received.
+func (h *Handler) handlePartialHistory(_ *service.Record, request *service.PartialHistoryRequest) {
+	if h.settings.GetXShared().GetValue() {
+		h.handlePartialHistoryAsync(request)
+	} else {
+		h.handlePartialHistorySync(request)
+	}
+}
+
+// handlePartialHistoryAsync handles a partial history request asynchronously. It is responsible for
+// collecting the history items until a full history record is received. This is the asynchronous version of handlePartialHistory.
+func (h *Handler) handlePartialHistoryAsync(request *service.PartialHistoryRequest) {
+	// This is the first partial history record we receive
+	if h.activeHistory == nil {
+		h.activeHistory = NewActiveHistory(
+			WithFlush(
+				func(_ *service.HistoryStep, items []*service.HistoryItem) {
+					record := &service.HistoryRecord{
+						Item: items,
+					}
+					h.flushHistory(record)
+				},
+			),
+		)
+	}
+	// Append the history items from the request to the current history record.
+	h.activeHistory.UpdateValues(request.Item)
+
+	// Flush the history record and start to collect a new one
+	if request.GetAction() == nil || request.GetAction().GetFlush() {
+		h.activeHistory.Flush()
+	}
+}
+
+// handlePartialHistorySync handles a partial history request. Collects the history items until a full
+// history record is received. This is the synchronous version of handlePartialHistory.
+func (h *Handler) handlePartialHistorySync(request *service.PartialHistoryRequest) {
+
+	// This is the first partial history record we receive
+	// for this step, so we need to initialize the history record
+	// and step. If the user provided a step in the request,
+	// use that, otherwise use 0.
+	if h.activeHistory == nil {
+		var step int64
+		switch {
+		case request.Step != nil:
+			step = request.Step.Num
+		case h.runRecord != nil:
+			step = h.runRecord.StartingStep
+		default:
+			// TODO: this should never happen, but we should handle it
+			// gracefully, should we raise an error?
+			step = 0
+		}
+
+		h.activeHistory = NewActiveHistory(
+			WithStep(step),
+			WithFlush(
+				func(step *service.HistoryStep, items []*service.HistoryItem) {
+					record := &service.HistoryRecord{
+						Step: step,
+						Item: items,
+					}
+					h.flushHistory(record)
+				},
+			),
+		)
+	}
+
+	// The HistoryRecord struct is responsible for tracking data related to
+	//	a single step in the history. Users can send multiple partial history
+	//	records for a single step. Each partial history record contains a
+	//	step number, a flush flag, and a list of history items.
+	//
+	// The step number indicates the step number for the history record. The
+	// flush flag determines whether the history record should be flushed
+	// after processing the request. The history items are appended to the
+	// existing history record.
+	//
+	// The following logic is used to process the request:
+	//
+	// -  If the request includes a step number and the step number is greater
+	//		than the current step number, the current history record is flushed
+	//		and a new history record is created.
+	// - If the step number in the request is less than the current step number,
+	//		we ignore the request and log a warning.
+	// 		NOTE: the server requires the steps of the history records
+	// 		to be monotonically increasing.
+	// -  If the step number in the request matches the current step number, the
+	//		history items are appended to the current history record.
+	//
+	// - If the request has a flush flag, another flush might occur after for the
+	// current history record after processing the request.
+	//
+	// - If the request doesn't have a step, and doesn't have a flush flag, this is
+	//	equivalent to step being equal to the current step number and a flush flag
+	//	being set to true.
+	if request.GetStep() != nil {
+		step := request.Step.GetNum()
+		current := h.activeHistory.GetStep().Num
+		if step > current {
+			h.activeHistory.Flush()
+			h.activeHistory.UpdateStep(step)
+		} else if step < current {
+			h.logger.CaptureWarn("received history record for a step that has already been received",
+				"received", step, "current", current)
+			return
+		}
+	}
+
+	// Append the history items from the request to the current history record.
+	h.activeHistory.UpdateValues(request.Item)
+
+	// Flush the history record and start to collect a new one with
+	// the next step number.
+	if (request.Step == nil && request.Action == nil) || (request.Action != nil && request.Action.Flush) {
+		h.activeHistory.Flush()
+		step := h.activeHistory.GetStep().Num + 1
+		h.activeHistory.UpdateStep(step)
+	}
+}
+
+func (h *Handler) handleSampledHistory(record *service.Record, response *service.Response) {
+	if h.sampledHistory == nil {
+		return
+	}
+	var items []*service.SampledHistoryItem
+
+	for key, sampler := range h.sampledHistory {
+		values := sampler.GetSample()
+		item := &service.SampledHistoryItem{
+			Key:         key,
+			ValuesFloat: values,
+		}
+		items = append(items, item)
+	}
+
+	response.ResponseType = &service.Response_SampledHistoryResponse{
+		SampledHistoryResponse: &service.SampledHistoryResponse{
+			Item: items,
+		},
+	}
+}
+
+// handleStepMetric handles the step metric for a given metric key. If the step metric is not
+// defined, it will be added to the defined metrics map.
+func (h *Handler) handleStepMetric(key string) {
+	if key == "" {
+		return
+	}
+
+	// already exists no need to add
+	if _, defined := h.metricHandler.definedMetrics[key]; defined {
+		return
+	}
+
+	metric, err := addMetric(key, key, &h.metricHandler.definedMetrics)
+
+	if err != nil {
+		h.logger.CaptureError("error adding metric to map", err)
+		return
+	}
+
+	stepRecord := &service.Record{
+		RecordType: &service.Record_Metric{
+			Metric: metric,
+		},
+		Control: &service.Control{
+			Local: true,
+		},
+	}
+	h.sendRecord(stepRecord)
+}
+
+func (h *Handler) handleMetric(record *service.Record, metric *service.MetricRecord) {
+	// metric can have a glob name or a name
+	// TODO: replace glob-name/name with one-of field
+	switch {
+	case metric.GetGlobName() != "":
+		if _, err := addMetric(metric, metric.GetGlobName(), &h.metricHandler.globMetrics); err != nil {
+			h.logger.CaptureError("error adding metric to map", err)
+			return
+		}
+		h.sendRecord(record)
+	case metric.GetName() != "":
+		if _, err := addMetric(metric, metric.GetName(), &h.metricHandler.definedMetrics); err != nil {
+			h.logger.CaptureError("error adding metric to map", err)
+			return
+		}
+		h.handleStepMetric(metric.GetStepMetric())
+		h.sendRecord(record)
+	default:
+		h.logger.CaptureError("invalid metric", errors.New("invalid metric"))
+	}
 }
