+++ core/pkg/server/metric.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"path/filepath"
 
-	"github.com/wandb/wandb/core/internal/corelib"
 	"github.com/wandb/wandb/core/pkg/service"
 	"google.golang.org/protobuf/proto"
 )
@@ -68,58 +67,6 @@ func (mh *MetricHandler) createMatchingGlobMetric(key string) *service.MetricRec
 	return nil
 }
 
-// handleStepMetric handles the step metric for a given metric key. If the step metric is not
-// defined, it will be added to the defined metrics map.
-func (h *Handler) handleStepMetric(key string) {
-	if key == "" {
-		return
-	}
-
-	// already exists no need to add
-	if _, defined := h.metricHandler.definedMetrics[key]; defined {
-		return
-	}
-
-	metric, err := addMetric(key, key, &h.metricHandler.definedMetrics)
-
-	if err != nil {
-		h.logger.CaptureError("error adding metric to map", err)
-		return
-	}
-
-	stepRecord := &service.Record{
-		RecordType: &service.Record_Metric{
-			Metric: metric,
-		},
-		Control: &service.Control{
-			Local: true,
-		},
-	}
-	h.sendRecord(stepRecord)
-}
-
-func (h *Handler) handleMetric(record *service.Record, metric *service.MetricRecord) {
-	// metric can have a glob name or a name
-	// TODO: replace glob-name/name with one-of field
-	switch {
-	case metric.GetGlobName() != "":
-		if _, err := addMetric(metric, metric.GetGlobName(), &h.metricHandler.globMetrics); err != nil {
-			h.logger.CaptureError("error adding metric to map", err)
-			return
-		}
-		h.sendRecord(record)
-	case metric.GetName() != "":
-		if _, err := addMetric(metric, metric.GetName(), &h.metricHandler.definedMetrics); err != nil {
-			h.logger.CaptureError("error adding metric to map", err)
-			return
-		}
-		h.handleStepMetric(metric.GetStepMetric())
-		h.sendRecord(record)
-	default:
-		h.logger.CaptureError("invalid metric", errors.New("invalid metric"))
-	}
-}
-
 type MetricSender struct {
 	definedMetrics map[string]*service.MetricRecord
 	metricIndex    map[string]int32
@@ -131,33 +78,5 @@ func NewMetricSender() *MetricSender {
 		definedMetrics: make(map[string]*service.MetricRecord),
 		metricIndex:    make(map[string]int32),
 		configMetrics:  make([]map[int]interface{}, 0),
-	}
-}
-
-// encodeMetricHints encodes the metric hints for the given metric record. The metric hints
-// are used to configure the plots in the UI.
-func (s *Sender) encodeMetricHints(_ *service.Record, metric *service.MetricRecord) {
-
-	_, err := addMetric(metric, metric.GetName(), &s.metricSender.definedMetrics)
-	if err != nil {
-		return
-	}
-
-	if metric.GetStepMetric() != "" {
-		index, ok := s.metricSender.metricIndex[metric.GetStepMetric()]
-		if ok {
-			metric = proto.Clone(metric).(*service.MetricRecord)
-			metric.StepMetric = ""
-			metric.StepMetricIndex = index + 1
-		}
-	}
-
-	encodeMetric := corelib.ProtoEncodeToDict(metric)
-	if index, ok := s.metricSender.metricIndex[metric.GetName()]; ok {
-		s.metricSender.configMetrics[index] = encodeMetric
-	} else {
-		nextIndex := len(s.metricSender.configMetrics)
-		s.metricSender.configMetrics = append(s.metricSender.configMetrics, encodeMetric)
-		s.metricSender.metricIndex[metric.GetName()] = int32(nextIndex)
 	}
 }
